Module CC.Actions

module T = T

module Lit = Lit

type proof = proof

type proof_step = proof_step

module P : Sidekick_core.PROOF with type lit = Lit.t and type t = proof and type term = T.Term.t and type proof_step = proof_step

type t = theory_actions

An action handle. It is used by the congruence closure to perform the actions below. How it performs the actions is not specified and is solver-specific.

val proof : t -> proof

val raise_conflict : t -> Lit.t list -> proof_step -> 'a

raise_conflict acts c pr declares that c is a tautology of the theory of congruence. This does not return (it should raise an exception).

parameter pr
the proof of c being a tautology

val propagate : t -> Lit.t -> reason:(unit -> Lit.t list * proof_step) -> unit

propagate acts lit ~reason pr declares that reason() => lit is a tautology.

reason() should return a list of literals that are currently true.

lit should be a literal of interest (see CC_S.set_as_lit).

This function might never be called, a congruence closure has the right to not propagate and only trigger conflicts.